moves global DApp config to root from util as it is not a util
