+++ src/components/web3/config.ts
@@ -1,18 +1,11 @@
+import {CoreProposalVoteChoices, VoteChoices} from './types';
+
 /**
  * WEB3 CONFIG for TRIBUTE DAO
  */
 
-/**
- * These addresses are important as the contracts use them in their configs.
- *
- * @todo Remove and use future npm package to get these values for tribute/laoland/moloch v3
- *
- * @see https://github.com/openlawteam/laoland/blob/9e0e03616a00e41e666351e146ee109b9fe37fb2/utils/DaoFactory.js
- */
-
-export const GUILD_ADDRESS = '0x000000000000000000000000000000000000dead';
-export const TOTAL_ADDRESS = '0x000000000000000000000000000000000000babe';
-export const SHARES_ADDRESS = '0x00000000000000000000000000000000000FF1CE';
-export const LOOT_ADDRESS = '0x00000000000000000000000000000000B105F00D';
-export const ETH_TOKEN_ADDRESS = '0x0000000000000000000000000000000000000000';
-export const DAI_TOKEN_ADDRESS = '0x95b58a6bff3d14b7db2f5cb5f0ad413dc2940658';
+// Vote choices should be "yes", "no" unless Moloch v3 contracts change.
+export const VOTE_CHOICES: CoreProposalVoteChoices = [
+  VoteChoices.yes,
+  VoteChoices.no,
+];
